Created Project component with individualized state in order to fix scoping problems on render prop
